+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+2.15.0 Release notes (YYYY-MM-DD)
+=============================================================
+### Compatibility
+* Sync protocol: 24
+* Server-side history format: 4
+* File format: 7
+* Realm Object Server: 3.0.0 or later
+
+### Breaking changes
+* None.
+
+### Enhancements
+* [Sync] Added `Realm.Sync.ConnectionState` representing the states a connection to the server can be in.
+* [Sync] Added support for `Session.addConnectionNotification()` and `Session.removeConnectionNotification`.
+* [Sync] Added `Session.connectionState`.
+* [Sync] Added `Session.isConnected()`.
+
+### Bug fixes
+* None
+
+### Internals
+* Updated to Object Store commit: 97fd03819f398b3c81c8b007feaca8636629050b
+
+
 2.14.2 Release notes (2018-8-8)
 =============================================================
 ### Compatibility